+++ src/messages/mod.rs
@@ -25,11 +25,10 @@ pub use protocol::{
     error_message, error_response, error_response_terminal, flush,
     insert_close_complete_after_last_close_complete, insert_close_complete_before_ready_for_query,
     insert_parse_complete_before_bind_complete, insert_parse_complete_before_parameter_description,
-    md5_challenge, md5_hash_password,
-    md5_hash_second_pass, md5_password, md5_password_with_hash, notify, parse_complete,
-    parse_params, parse_startup, plain_password_challenge, read_password, ready_for_query,
-    scram_server_response, scram_start_challenge, server_parameter_message, simple_query,
-    ssl_request, startup, sync, wrong_password,
+    md5_challenge, md5_hash_password, md5_hash_second_pass, md5_password, md5_password_with_hash,
+    notify, parse_complete, parse_params, parse_startup, plain_password_challenge, read_password,
+    ready_for_query, scram_server_response, scram_start_challenge, server_parameter_message,
+    simple_query, ssl_request, startup, sync, wrong_password,
 };
 pub use socket::{
     proxy_copy_data, proxy_copy_data_with_timeout, read_message, read_message_data,